# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CRT Portal API",
    "description": "API documentation for CRT Portal",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/students/{id}": {
      "get": {
        "tags": [
          "student-controller"
        ],
        "operationId": "getStudentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "student-controller"
        ],
        "operationId": "updateStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "student-controller"
        ],
        "operationId": "deleteStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sections/{sectionId}": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "operationId": "getSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "section-controller"
        ],
        "operationId": "updateSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "section-controller"
        ],
        "operationId": "deleteSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sections/student/{studentId}/section/{sectionId}": {
      "put": {
        "tags": [
          "section-controller"
        ],
        "operationId": "updateStudentSection",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/students": {
      "get": {
        "tags": [
          "student-controller"
        ],
        "operationId": "getAllStudents",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "student-controller"
        ],
        "operationId": "createStudent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sections": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "operationId": "getAllSections",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "section-controller"
        ],
        "operationId": "createSection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/sections/{sectionId}/students": {
      "post": {
        "tags": [
          "section-controller"
        ],
        "operationId": "registerStudents",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/time-slots/section/{sectionId}": {
      "get": {
        "tags": [
          "time-slot-controller"
        ],
        "operationId": "getTimeSlotsBySection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimeSlotDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sections/trainer/{trainerId}": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "operationId": "getSectionsByTrainer",
        "parameters": [
          {
            "name": "trainerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/section-schedules/section/{sectionId}": {
      "get": {
        "tags": [
          "section-schedule-controller"
        ],
        "operationId": "getScheduleBySection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionScheduleDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/attendance/section/{sectionId}": {
      "get": {
        "tags": [
          "attendance-controller"
        ],
        "operationId": "getSectionAttendanceRecords",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionAttendanceRecordDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TimeSlotDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "breakDescription": {
            "type": "string"
          },
          "inchargeFacultyId": {
            "type": "string",
            "format": "uuid"
          },
          "sectionId": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "break": {
            "type": "boolean"
          }
        }
      },
      "StudentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "rollNumber": {
            "type": "string"
          },
          "regNum": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "batch": {
            "type": "string",
            "enum": [
              "Y22",
              "Y23",
              "Y24",
              "Y25"
            ]
          },
          "crtEligibility": {
            "type": "boolean"
          },
          "feedback": {
            "type": "string"
          },
          "attendancePercentage": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CreateSectionDTO": {
        "type": "object",
        "properties": {
          "trainerId": {
            "type": "string",
            "format": "uuid"
          },
          "sectionName": {
            "type": "string"
          }
        }
      },
      "CRT_TrainerDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "SectionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "trainer": {
            "$ref": "#/components/schemas/CRT_TrainerDTO"
          },
          "students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentDTO"
            },
            "uniqueItems": true
          },
          "strength": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SectionScheduleDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sectionId": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "timeSlots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeSlotDTO"
            }
          }
        }
      },
...
      "SectionAttendanceRecordDTO": {
        "type": "object",
        "properties": {
          "regNum": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "attendancePercentage": {
            "type": "number",
            "format": "double"
          },
          "monthTitle": {
            "type": "string"
          },
          "totalClasses": {
            "type": "integer",
            "format": "int64"
          },
          "absences": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}